rm memcmp tracer example & added simple example
