+++ app/tracer/example/Enclave/encl.c
@@ -1,0 +1,7 @@
+#include <stdint.h>
+#include <string.h>
+
+int dummy_enclave_fn()
+{
+    return 42;
+}
