--- app/tracer/memcmp/.gitignore
@@ -1,17 +1,0 @@
-app
-measurements.txt
-measurements_raw.txt
-outlier_idx.txt
-plot.pdf
-xlabels.gp
-
-*.swp
-
-pages_dbg.txt
-pages_dbg_bit.txt
-*.vcd
-
-out.txt
-parsed_nop.txt
-parsed_zz.txt
-parsed_strlen.txt
