--- app/tracer/memcmp/Enclave/encl.edl
@@ -1,9 +1,0 @@
-enclave {
-	
-	trusted {
-        public int memcmp_pwd([in,string] char *pwd);
-	};
-	
-	untrusted {
-	};
-};
